refactor: reduce createSet args in const-specific fns
