+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/testutils/commonPlusTests.kt
@@ -1,6 +1,7 @@
 package xyz.lbres.kotlinutils.set.multiset.testutils
 
 import xyz.lbres.kotlinutils.assertEmpty
+import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSet
@@ -12,9 +13,10 @@ private val e1 = ArithmeticException()
 private val e2 = NullPointerException()
 private val e3 = IllegalArgumentException()
 
+@Suppress(Suppressions.UNCHECKED_CAST)
 fun runPlusTests(
     createSet: (Collection<*>) -> MultiSet<*>,
-    createOtherSet: (Collection<*>) -> MultiSet<*>,
+    createOtherSet: (Collection<*>) -> MultiSet<*>
 ) {
     val createIntSet: (Collection<Int>) -> MultiSet<Int> = { createSet(it) as MultiSet<Int> }
     val createStringSet: (Collection<String>) -> MultiSet<String> = { createSet(it) as MultiSet<String> }
